Optimization practice
Reduce branch mis-predicts by 10-15%. Most of it was from re-working
err_check macro to only branch on error, not on normal operation. This
allows the 'back not forwards' (assume the forward jump is not taken,
but a backwards one is taken) prediction to be correct. Also re-work
some other comparisons on the same principle. Replace a loop with 'repe scasb'.

Reduce code size by favoring 32-bit registers. Those seem to encode
at least as small, sometimes smaller, than 16 and 8 bit registers.